http2: end stream eagerly after sending the request body
Check EOF eagerly on the request body when its content-length
is specified and it is expected to end. Thus, the data frame
containing the last chunk of data of the body will be marked with
END_STREAM eagerly.

However, if the specified content-length is incorrect, the whole
request body will still be sent, regardless of its actual length.
This will render the request malformed according to the spec.
We believe it is the receiving peer's responsibility to handle
the error.

Fix golang/go#32254

Change-Id: Id24c043c7cc3a41421dfd099a139f1b1e08056b9